[FIX] Complete "Suggested Distro" section Suggested Distro section is now (finally) completed: 1) Definetely shortened. There is no more subsections (Ubuntu vs Arch). The only two suggested distros are Manjaro and Ubuntu flavors 2) Added bold distro names to improve readabilty 3) minor fixes (typos here and there)
The reason behind this drastic change is aim of the guide: it must
be focused on distro optimization, not distro choice! Moreover,
experienced users don't need someone to advice them, since
choosing a distro is like choosing a girlfriend... cmon you wanna
do it yourself! For new users,  Ubuntu and Manjaro are always good
suggestions.